Create Tui (thats what I meant with "auslagern"), finish TuiSpec make the main class test disappear
