+++ src/main/scala/Schach/Tui.scala
@@ -1,0 +1,38 @@
+package Schach
+
+class Tui {
+
+  def readInput(line: String): Array[Int] = {
+    val fromX = getPoint(line.charAt(0))
+    val fromY = getPoint(line.charAt(1))
+    val toX = getPoint(line.charAt(3))
+    val toY = getPoint(line.charAt(4))
+    Array(fromX, fromY, toX, toY)
+  }
+
+  def controlInput(line: String): Boolean = {
+    line.matches("([A-H][1-8]\\s[A-H][1-8])")
+  }
+
+  def getPoint(input: Char): Int = {
+    input match {
+      case 'A' => 0
+      case 'B' => 1
+      case 'C' => 2
+      case 'D' => 3
+      case 'E' => 4
+      case 'F' => 5
+      case 'G' => 6
+      case 'H' => 7
+      case '1' => 0
+      case '2' => 1
+      case '3' => 2
+      case '4' => 3
+      case '5' => 4
+      case '6' => 5
+      case '7' => 6
+      case '8' => 7
+      case _ => -1
+    }
+  }
+}
